refactor: Code cleanup (splitted classes, added docblocks...)
Closes Taiga #84 #78
